+++ Microsoft.Azure.Cosmos/src/SqlObjects/Visitors/SqlObjectEqualityVisitor.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Cosmos.SqlObjects.Visitors
                 return false;
             }
 
-            if (!first.Alias.Accept(this, second.Alias))
+            if (!SqlObjectEqualityVisitor.Equals(first.Alias, second.Alias))
             {
                 return false;
             }
@@ -308,7 +308,7 @@ namespace Microsoft.Azure.Cosmos.SqlObjects.Visitors
                 return false;
             }
 
-            if (!first.ParentPath.Accept(this, second.ParentPath))
+            if (!SqlObjectEqualityVisitor.Equals(first.ParentPath, second.ParentPath))
             {
                 return false;
             }
@@ -328,7 +328,7 @@ namespace Microsoft.Azure.Cosmos.SqlObjects.Visitors
                 return false;
             }
 
-            if (!first.RelativePath.Accept(this, second.RelativePath))
+            if (!SqlObjectEqualityVisitor.Equals(first.RelativePath, second.RelativePath))
             {
                 return false;
             }
